+++ src/client/js/functions.js
@@ -1,22 +1,12 @@
 /*
  * pwix:accounts-ui/src/client/js/functions.js
+ *
+ * This is the external API on client-side.
  */
 
 import _ from 'lodash';
 
 import { Tracker } from 'meteor/tracker';
-
-_dropdownItems = {
-    dep: null,
-    state: null,
-    items: []
-};
-
-_dropdownItemsExt = {
-    dep: null,
-    state: null,
-    items: []
-};
 
 AccountsUI = {
     ...AccountsUI,
@@ -35,50 +25,6 @@ AccountsUI = {
                     instance.$( '.ac-panel' ).trigger( 'ac-clear-panel' );
                 }
             }
-        },
-
-        /**
-         * @locus Client
-         * @returns {Array} the list of dropdown items to be displayed regarding the
-         *  current user connection state, as an array of '<a>...</a>' strings
-         *  NB: we only return here the list of standard dropdown items as we cannot manage
-         *  any acUserLogin configuration at this global level.
-         *  A reactive data source.
-         */
-        dropdownItems(){
-            if( !_dropdownItems.dep ){
-                _dropdownItems.dep = new Tracker.Dependency();
-                _dropdownItems.dep.depend();
-            }
-            const state = AccountsUI.Connection.state();
-            if( state !== _dropdownItems.state ){
-                _dropdownItems.state = state;
-                _dropdownItems.items = _buildStandardItems( AccountsUI.dropdownItemsExt());
-                _dropdownItems.dep.changed();
-            }
-            return _dropdownItems.items;
-        },
-
-        /**
-         * @locus Client
-         * @returns {Array} the list of dropdown items to be displayed regarding the
-         *  current user connection state.
-         *  NB: we only return here the list of standard dropdown items as we cannot manage
-         *  any acUserLogin configuration at this global level.
-         *  A reactive data source.
-         */
-        dropdownItemsExt(){
-            if( !_dropdownItemsExt.dep ){
-                _dropdownItemsExt.dep = new Tracker.Dependency();
-                _dropdownItemsExt.dep.depend();
-            }
-            const state = AccountsUI.Connection.state();
-            if( state !== _dropdownItemsExt.state ){
-                _dropdownItemsExt.state = state;
-                _dropdownItemsExt.items = _stdMenuItems[state];
-                _dropdownItemsExt.dep.changed();
-            }
-            return _dropdownItemsExt.items;
         },
 
         /**